[api] Turn global JS files into TS ones
